modified:   Makefile 	modified:   src/apple.h 	modified:   src/config.h 	modified:   src/main.cpp 	modified:   src/screen.h 	modified:   src/snake.h

+++ src/config.h
@@ -1,9 +1,9 @@
 #pragma once
 /* -----CONFIG FILE----- */
-inline const unsigned short     SCREEN_WIDTH = 864;
-inline const unsigned short     SCREEN_LENGHT = 672;
-inline const unsigned short     TEXTURE_SIZE = 48;
-inline const unsigned short     BORDER_SIZE = 48;
-inline const unsigned short     FPS = 75;
+inline const Uint16             SCREEN_WIDTH = 864;
+inline const Uint16             SCREEN_LENGHT = 672;
+inline const Uint8              TEXTURE_SIZE = 48;
+inline const Uint8              BORDER_SIZE = 48;
+inline const Uint8              FPS = 75;
 inline const float              SNAKE_SPEED = 2.0f;
 inline const char*              TITLE = "Cnake";
